+++ R/scale_longitude.R
@@ -26,7 +26,7 @@
 #' @import ggplot2
 scale_x_longitude <- function(ticks = 60, name = "", expand = c(0, 0),
                               breaks = seq(0, 360 - ticks, by = ticks),
-                              labels = ConvertLongitude(breaks, 360),
+                              labels = LonLabel(breaks),
                               ...) {
     ggplot2::scale_x_continuous(name = name, expand = expand,
                                 breaks = breaks,
@@ -39,7 +39,7 @@ scale_x_longitude <- function(ticks = 60, name = "", expand = c(0, 0),
 #' @import ggplot2
 scale_y_longitude <- function(ticks = 60, name = "", expand = c(0, 0),
                               breaks = seq(0, 360 - ticks, by = ticks),
-                              labels = ConvertLongitude(breaks, 360),
+                              labels = LonLabel(breaks),
                               ...) {
     ggplot2::scale_y_continuous(name = name, expand = expand,
                        breaks = breaks,
@@ -52,9 +52,10 @@ scale_y_longitude <- function(ticks = 60, name = "", expand = c(0, 0),
 #' @export
 #' @import ggplot2
 scale_x_latitude <- function(ticks = 30, name = "", expand = c(0, 0),
-                             breaks = seq(-90, 90, by = ticks), ...) {
+                             breaks = seq(-90, 90, by = ticks),
+                             labels = LatLabel(breaks), ...) {
     ggplot2::scale_x_continuous(name = name, expand = expand,
-                                breaks = breaks,
+                                breaks = breaks, labels = labels,
                                 ...)
 }
 
@@ -62,9 +63,10 @@ scale_x_latitude <- function(ticks = 30, name = "", expand = c(0, 0),
 #' @export
 #' @import ggplot2
 scale_y_latitude <- function(ticks = 30, name = "", expand = c(0, 0),
-                             breaks = seq(-90, 90, by = ticks), ...) {
+                             breaks = seq(-90, 90, by = ticks),
+                             labels = LatLabel(breaks), ...) {
     ggplot2::scale_y_continuous(name = name, expand = expand,
-                       breaks = breaks,
+                       breaks = breaks, labels = labels,
                        ...)
 }
 
@@ -84,4 +86,45 @@ scale_y_level <- function(name = "", expand = c(0, 0), trans = "reverselog", ...
     ggplot2::scale_y_continuous(name = name, expand = expand,
                        trans = trans,
                        ...)
+}
+
+
+
+#' Label longitude and latitude
+#'
+#' Provide easy functions for adding suffixes to longitude and latitude for labeling
+#' maps.
+#'
+#' @param lon longitude in degrees between 0 and 360
+#' @param lat latitude in degrees
+#' @param east,west,north,south text to append for each cuadrant
+#'
+#' @details
+#' The default values are for Spanish.
+#'
+#' @examples
+#'
+#' LonLabel(0:360)
+#'
+#' # If you want it to speak in English.
+#' LonLabel <- function(...) metR::LonLabel(..., west = "\u00B0W")
+#'
+#'
+#' @export
+#' @name map_labels
+#' @family ggplot2 helpers
+LonLabel <- function(lon, east = "\u00B0E", west = "\u00B0O") {
+    lon <- ConvertLongitude(lon)
+    newlon <- ifelse(lon < 0, paste0(abs(lon), west), paste0(lon, east))
+    newlon[lon == 0] <- paste0(lon[lon == 0], "\u00B0")
+    newlon[lon == 180] <- paste0(lon[lon == 180], "\u00B0")
+    return(newlon)
+}
+
+#' @export
+#' @rdname map_labels
+LatLabel <- function(lat, north = "\u00B0N", south = "\u00B0S") {
+    newlat <- ifelse(lat < 0, paste0(abs(lat), south), paste0(lat, north))
+    newlat[lat == 0] <- paste0(lat[lat == 0], "\u00B0")
+    return(newlat)
 }